Move php-cs-fixer utils to own namespace
Extract common helpers to accessible namespace to allow shared access for upcoming PHP_CodeSniffer integration.
