+++ lib/Extension/LanguageServerPhpCsFixer/Provider/PhpCsFixerDiagnosticsProvider.php
@@ -5,8 +5,8 @@ namespace Phpactor\Extension\LanguageServerPhpCsFixer\Provider;
 use Amp\CancellationToken;
 use Amp\Promise;
 use Amp\Success;
+use Phpactor\Diff\StringSharedChars;
 use Phpactor\Extension\LanguageServerPhpCsFixer\Model\PhpCsFixerProcess;
-use Phpactor\Extension\LanguageServerPhpCsFixer\Util\StringSharedChars;
 use Phpactor\LanguageServerProtocol\CodeAction;
 use Phpactor\LanguageServerProtocol\Command;
 use Phpactor\LanguageServerProtocol\Diagnostic;
